Change type of game id to int

+++ vaulthiveserver/src/main/java/dev/septian/vaulthiveserver/domain/Game.java
@@ -1,9 +1,6 @@
 package dev.septian.vaulthiveserver.domain;
 
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonRawValue;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,8 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Game {
-
-    private long id;
+    private int id;
     private String name;
     private String description;
     private String released;
